TowardsMalmoIntegration
Malmo is called from MalmoEvaluator.
Video can be created for every case.
The files can be saved in a particular folder (so that not every Malmo
instance overwrites the results).
PROBLEM: The anti-virus identifies as malware this version! (It creates
a list of compressed, that is, encoded, folders... similarly to
ransomware?)